New manufactured functions to verify PML solutions and rearrangement of MAIN menu
Inside module bessel_evaluation.F90 there was added the rescaled Bessel-Frobenius algorithm with quadruple precision (bessel_new), with a preset solution of the bent three-layer waveguide based on bessel_new, with preset eigenvalues and coefficients computed with Julia high precision eigensolver. Subroutine set_initial_mesh also revised to have consistent boundary conditions with the eigensolver setup.

The PML functions were updated, as we now require to evaluate functions at the complex stretched variables and not only the Jacobians.

The main file interactive menu was updated to present a better organization of the refinements and the possibility of multiple uniform refinements + solve + error/residual computation, which was available for anisotropic only.